Refactor: move agreement dialog & confirm buttons to `UploadAddressesWidget`
