+++ src/modules/dashboard/components/Whitelist/UploadAddressesWidget/UploadAddressesWidget.css.d.ts
@@ -4,3 +4,5 @@ export const inputContainer: string;
 export const input: string;
 export const actionsContainer: string;
 export const actionsSubContainer: string;
+export const buttonsContainer: string;
+export const agreeemntButton: string;
