--- src/modules/dashboard/components/Whitelist/Whitelist.css
@@ -1,9 +1,0 @@
-.buttonsContainer {
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-}
-
-.agreeemntButton {
-  margin-right: 30px;
-}
