--- src/modules/dashboard/components/Whitelist/Whitelist.css.d.ts
@@ -1,2 +1,0 @@
-export const buttonsContainer: string;
-export const agreeemntButton: string;
